requirements.txt
appdirs==1.4.3
asgiref==3.2.3
Babel==2.7.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.8.1
djangorestframework==3.10.2
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.5.0
psycopg2==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
virtualenv==20.0.4
whitenoise==5.0.1
zipp==2.2.0
